amd: move ISARenderer into renderer/isa/amd.py (PR1 skeleton)
Relocate out of runtime/autogen. Drop TC_LDS staging, WMMA schedule
hoist/sink, and tensor_cores advertising. ALU/global/LDS/waitcnt/ELF
remain. WMMA tests skipped until PR3.
