+++ test/amd/test_amd_renderer.py
@@ -13,7 +13,7 @@ from tinygrad.runtime.autogen import amdgpu_kd
 from tinygrad.runtime.support.elf import elf_loader
 from tinygrad.uop import Ops
 from tinygrad.uop.ops import AxisType, KernelInfo, UOp, graph_rewrite
-import tinygrad.runtime.autogen.amd.rdna3.amd_lib as amd_lib
+import tinygrad.renderer.isa.amd as amd_lib
 from tinygrad.renderer.isa.amd import AMDRenderer, AMDOps
 
 _GFX11 = Target("AMD", arch="gfx1100")
@@ -712,6 +712,7 @@ class TestAMDRenderer(unittest.TestCase):
     self.assertTrue(renderer.supports_float4)
     self.assertEqual(renderer.local_prod_max, 1024)
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_tensor_cores_only_advertise_lowered_mode(self):
     self.assertEqual([(x.dtype_in, x.dtype_out) for x in _REN.tensor_cores], [(dtypes.half, dtypes.float)])
 
@@ -969,6 +970,7 @@ class TestAMDRenderer(unittest.TestCase):
     self.assertEqual(inst_names.count("GLOBAL_STORE_B128"), 4)
     self.assertEqual(inst_names.count("V_FMA_F32"), 48)
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_wmma_assembles(self):
     prg = _half_matmul_wmma_program()
     # Small WMMA may still v_pack; product-16 register path uses u16+d16_hi (tested below).
@@ -1041,14 +1043,17 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_tc_lds_ab_stages_without_spill(self):
     # Dual LOCAL 2×2 WG + tile UPCAST 2×2 (product ≤4); default bounce.
     self._half_matmul_tc_lds_ab_pre_regalloc()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_tc_lds_ab_local4_stages_without_spill(self):
     # TC_LOCAL=4 → asymmetric LOCAL 4×2 (symmetric 4×4 historically TDR on display).
     self._half_matmul_tc_lds_ab_pre_regalloc(tc_local=4)
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_tc_lds_tid_fill_peer_lds_reads(self):
     # Opt-in tid-fill (TC_LDS_TID=1): peer remap → DS_LOAD_U16 when A block==threads
     # (needs A tile_prod=local_n*2). Default LDS tiles 2×2 → tile_prod=2 → bounce fallback.
@@ -1080,6 +1085,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_tc_lds_bounce_no_u16_lds(self):
     # Hybrid bounce: shared A (tid-fill/major-read) + B tid-wide → GLOBAL B128, no peer U16.
     # A LDS is block×16 (=4KiB at LOCAL 2×2); total A+B ≤ 8KiB.
@@ -1110,6 +1116,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_tc_lds_ab_stages_without_spill_gate(self):
     # Default LDS: product 8 + B transpose + addr remat → spill-free, DS_LOAD_B128 on A/B.
     import os
@@ -1143,6 +1150,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_upcast4_without_spill(self):
     # Full 4×4 (16 WMMA) is ALLOW_UPCAST16-only. Even SPILL=0 offline can TDR display GPUs
     # (vgpr_span≈254). Register path; expect ≥16 WMMA. Scratch may be high — do not ship.
@@ -1177,6 +1185,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_lds_upcast16_aliases_acc_not_reg_scratch(self):
     # LDS product-16 uses MOV-zero cin (vs register SLOAD reload). ACC alias must still
     # kill the 128-slot REG buffer SLOAD/SSTORE. Residual VGPR spills OK — not default yet.
@@ -1205,6 +1214,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_upcast16_refused_without_allow(self):
     # Register defaults to product 16; ALLOW_UPCAST16=0 clamps back to ≤8.
     import os
@@ -1232,6 +1242,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_default_lds_intensity_gates(self):
     # Opt-in LDS defaults to product 8, spill-free with EXTRACT+addr remat.
     import os
@@ -1274,6 +1285,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_default_is_spill_free_sixteen_wmma(self):
     # Default ISA: register path, product 16 → 16 WMMA, spill-free (LDS is opt-in).
     import os
@@ -1300,6 +1312,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_c_stores_share_peeled_base(self):
     # Soft-peel nested ADD+imm so 128 C stores share one addr base; large byte offs use v_lshl_add.
     import os
@@ -1418,6 +1431,7 @@ class TestAMDRenderer(unittest.TestCase):
     self.assertNotIn("GLOBAL_LOAD_U16", inst_names)
     self.assertNotIn("GLOBAL_STORE_B16", inst_names)
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_uses_wide_global_load(self):
     # Register path: contiguous operand B128; strided WMMA operand stays U16.
     import os
@@ -1442,6 +1456,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_register_path_strided_operand_is_u16(self):
     # Contiguous A → B128×8; strided B → u16+d16_hi into pack VGPR (default).
     # TC_LOCAL=4 (register default) + d16_hi; opt out with AMD_D16_HI=0 / TC_LOCAL=0.
@@ -1473,6 +1488,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_register_path_vpack_opt_out(self):
     # AMD_D16_HI=0 + TC_LOCAL=0 → legacy u16×64 + v_pack (no LOCAL).
     import os
@@ -1500,6 +1516,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_tc_lds_ab_not_default_until_faster(self):
     # Step C: ISA default is register+B128 until LDS wins @4096 on HW.
     import os
@@ -1520,6 +1537,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_lds_product8_clamped_without_allow(self):
     # ALLOW_LDS_PRODUCT8=0 forces 2×2 under LDS (product 8 is otherwise the LDS default).
     import os
@@ -1544,6 +1562,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_lds_product8_allow_serializes_a_batches(self):
     # ALLOW_LDS_PRODUCT8: expand serializes A batches (AFTER). Spill-free with LLOAD EXTRACT remat.
     import os
@@ -1569,6 +1588,7 @@ class TestAMDRenderer(unittest.TestCase):
       getenv.cache_clear()
       to_program_cache.clear()
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_lds_unroll_gated_when_over_tile_budget(self):
     # TC_LDS_UNROLL used to apply then crash in expand_wmma (tile product > budget).
     # With product-8 already at max_tiles, UNROLL must not apply — compile stays green.
@@ -1884,6 +1904,7 @@ class TestAMDRenderer(unittest.TestCase):
     # 1D locals: USER_SGPR stays 2 so WGID lands at s2 (hand kernels / LLVM).
     self.assertEqual((desc.compute_pgm_rsrc2 >> amdgpu_kd.COMPUTE_PGM_RSRC2_USER_SGPR_COUNT_SHIFT) & 0x1f, 2)
 
+  @unittest.skip("PR3 WMMA/TC")
   def test_half_matmul_2d_locals_pads_user_sgpr_on_gfx1100(self):
     # gfx1100 UserSGPRInit16Bug: 2D locals need USER_SGPR=15 so workitem Y is initialized.
     import os
